refactor(reply): move pipeline and reviewer into runtime package
