+++ docs/REPOSITORY_LAYOUT.md
@@ -7,7 +7,8 @@
 | 路径 | 职责 |
 | --- | --- |
 | `local_server.py` | 本机 HTTP 服务、信件主流程和后台媒体任务入口 |
-| `reply_pipeline.py`、`reply_*.py` | canonical reply 装配、审校、持久化和媒体投影 |
+| `reply_pipeline.py`、`reply_reviewer.py` | 原有导入兼容别名，映射到 `runtime/reply/` 中的同一模块对象 |
+| `runtime/reply/`、其他 `reply_*.py` | canonical reply 装配、审校、持久化和媒体投影 |
 | `llm_gateway.py`、`voice_direction.py` | 可配置模型调用与语音导演计划 |
 | `conversation_memory_*.py`、`mem0_memory.py` | 长期记忆、outbox、管理和运行时适配 |
 | `private_world_*.py` | 私有关系事件、ledger、reducer、投影和管理 |
