+++ tests/http/test_reply_delay_and_media.py
@@ -12,7 +12,7 @@ import local_server
 from letter_triage import TriageResult
 from reply_context import ReplyMode
 from reply_orchestrator import ReplyState
-from reply_pipeline import PipelineResult
+from runtime.reply.reply_pipeline import PipelineResult
 from voice_direction import VoiceDirectionError, VoicePerformancePlan
