+++ tests/persona/test_persona_v2_release_e2e.py
@@ -19,8 +19,8 @@ from private_world_port import ContinuationAwareness, PrivateWorldSnapshot
 from private_world_reducer import ReducerEventKind
 from reply_context import ReplyContext, ReplyMode, TrustedTime
 from reply_orchestrator import ReplyResult, ReplyState
-from reply_pipeline import ReplyPipeline, UnavailableRewriter
-from reply_reviewer import (
+from runtime.reply.reply_pipeline import ReplyPipeline, UnavailableRewriter
+from runtime.reply.reply_reviewer import (
     NullReviewer,
     ReviewResult,
     ReviewerScores,
